#1073 Allow nginx ingress for keptn api

+++ installer/scripts/common/setupKeptn.sh
@@ -13,39 +13,6 @@ wait_for_deployment_in_namespace "nats-operator" "keptn"
 kubectl apply -f ../manifests/nats/nats-cluster.yaml
 verify_kubectl $? "Creating NATS Cluster failed."
 
-# Domain used for routing to keptn services
-wait_for_istio_ingressgateway "hostname"
-DOMAIN=$(kubectl get svc istio-ingressgateway -o json -n istio-system | jq -r .status.loadBalancer.ingress[0].hostname)
-if [[ $? != 0 ]]; then
-  print_error "Failed to get ingress gateway information." && exit 1
-fi
-
-if [[ "$DOMAIN" == "null" && "$GATEWAY_TYPE" == "LoadBalancer" ]]; then
-  print_info "Could not get ingress gateway domain name. Trying to retrieve IP address instead."
-  
-  wait_for_istio_ingressgateway "ip"
-
-  DOMAIN=$(kubectl get svc istio-ingressgateway -o json -n istio-system | jq -r .status.loadBalancer.ingress[0].ip)
-  if [[ "$DOMAIN" == "null" ]]; then
-    print_error "IP of Istio Ingressgateway could not be derived."
-    exit 1
-  fi
-  DOMAIN="$DOMAIN.xip.io"
-elif [[ "$DOMAIN" == "null" && "$GATEWAY_TYPE" == "NodePort" ]]; then
-  NODE_PORT=$(kubectl -n istio-system get service istio-ingressgateway -o jsonpath='{.spec.ports[?(@.name=="https")].nodePort}')
-  NODE_IP=$(kubectl get nodes -l node-role.kubernetes.io/worker=true -o jsonpath='{ $.items[*].status.addresses[?(@.type=="InternalIP")].address }')
-  DOMAIN="$NODE_IP:$NODE_PORT"
-fi
-
-# Set up SSL
-openssl req -nodes -newkey rsa:2048 -keyout key.pem -out certificate.pem  -x509 -days 365 -subj "/CN=$DOMAIN"
-
-kubectl create --namespace istio-system secret tls istio-ingressgateway-certs --key key.pem --cert certificate.pem
-#verify_kubectl $? "Creating secret for istio-ingressgateway-certs failed."
-
-rm key.pem
-rm certificate.pem
-
 # Add config map in keptn namespace that contains the domain - this will be used by other services as well
 rm ../manifests/gen/keptn-domain-configmap.yaml
 
@@ -121,12 +88,14 @@ wait_for_deployment_in_namespace "lighthouse-service-start-evaluation-distributo
 wait_for_deployment_in_namespace "gatekeeper-service-evaluation-done-distributor" "keptn"
 wait_for_deployment_in_namespace "shipyard-service-create-project-distributor" "keptn"
 
-kubectl apply -f ../manifests/keptn/keptn-gateway.yaml
-verify_kubectl $? "Deploying keptn gateway failed."
+if [ "$INGRESS" = "istio" ]; then
+  kubectl apply -f ../manifests/keptn/keptn-gateway.yaml
+  verify_kubectl $? "Deploying keptn gateway failed."
 
-rm -f ../manifests/keptn/gen/keptn-api-virtualservice.yaml
-cat ../manifests/keptn/keptn-api-virtualservice.yaml | \
-  sed 's~DOMAIN_PLACEHOLDER~'"$DOMAIN"'~' > ../manifests/keptn/gen/keptn-api-virtualservice.yaml
+  rm -f ../manifests/keptn/gen/keptn-api-virtualservice.yaml
+  cat ../manifests/keptn/keptn-api-virtualservice.yaml | \
+    sed 's~DOMAIN_PLACEHOLDER~'"$DOMAIN"'~' > ../manifests/keptn/gen/keptn-api-virtualservice.yaml
 
-kubectl apply -f ../manifests/keptn/gen/keptn-api-virtualservice.yaml
-verify_kubectl $? "Deploying keptn api virtualservice failed."
+  kubectl apply -f ../manifests/keptn/gen/keptn-api-virtualservice.yaml
+  verify_kubectl $? "Deploying keptn api virtualservice failed."
+fi
