+++ installer/scripts/openshift/setupKeptn.sh
@@ -13,39 +13,6 @@ wait_for_deployment_in_namespace "nats-operator" "keptn"
 kubectl apply -f ../manifests/nats/nats-cluster.yaml
 verify_kubectl $? "Creating NATS Cluster failed."
 
-ROUTER_POD=$(oc get pods -n default -l router=router -ojsonpath={.items[0].metadata.name})
-# allow wildcard domains
-oc project default
-oc adm router --replicas=0
-verify_kubectl $? "Scaling down router failed"
-oc set env dc/router ROUTER_ALLOW_WILDCARD_ROUTES=true
-verify_kubectl $? "Configuration of openshift router failed"
-oc scale dc/router --replicas=1
-verify_kubectl $? "Upscaling of router failed"
-
-oc delete pod $ROUTER_POD -n default --force --grace-period=0 --ignore-not-found
-
-# create wildcard route for istio ingress gateway
-
-BASE_URL=$(oc get route -n istio-system istio-ingressgateway -oyaml | yq r - spec.host | sed 's~istio-ingressgateway-istio-system.~~')
-# Domain used for routing to keptn services
-DOMAIN="ingress-gateway.$BASE_URL"
-
-oc create route passthrough istio-wildcard-ingress-secure-keptn --service=istio-ingressgateway --hostname="www.keptn.ingress-gateway.$BASE_URL" --port=https --wildcard-policy=Subdomain --insecure-policy='None' -n istio-system
-
-
-oc adm policy  add-cluster-role-to-user cluster-admin system:serviceaccount:keptn:default
-verify_kubectl $? "Adding cluster-role failed."
-
-# Set up SSL
-openssl req -nodes -newkey rsa:2048 -keyout key.pem -out certificate.pem  -x509 -days 365 -subj "/CN=$DOMAIN"
-
-kubectl create --namespace istio-system secret tls istio-ingressgateway-certs --key key.pem --cert certificate.pem
-#verify_kubectl $? "Creating secret for istio-ingressgateway-certs failed."
-
-rm key.pem
-rm certificate.pem
-
 #verify_kubectl $? "Creation of keptn ingress route failed."
 
 
@@ -118,15 +85,17 @@ wait_for_deployment_in_namespace "shipyard-service" "keptn"
 wait_for_deployment_in_namespace "lighthouse-service" "keptn"
 wait_for_deployment_in_namespace "configuration-service" "keptn"
 
-kubectl apply -f ../manifests/keptn/keptn-gateway.yaml
-verify_kubectl $? "Deploying keptn gateway failed."
+if [ "$INGRESS" = "istio" ]; then
+  kubectl apply -f ../manifests/keptn/keptn-gateway.yaml
+  verify_kubectl $? "Deploying keptn gateway failed."
 
-rm -f ../manifests/keptn/gen/keptn-api-virtualservice.yaml
-cat ../manifests/keptn/keptn-api-virtualservice.yaml | \
-  sed 's~DOMAIN_PLACEHOLDER~'"$DOMAIN"'~' > ../manifests/keptn/gen/keptn-api-virtualservice.yaml
+  rm -f ../manifests/keptn/gen/keptn-api-virtualservice.yaml
+  cat ../manifests/keptn/keptn-api-virtualservice.yaml | \
+    sed 's~DOMAIN_PLACEHOLDER~'"$DOMAIN"'~' > ../manifests/keptn/gen/keptn-api-virtualservice.yaml
 
-kubectl apply -f ../manifests/keptn/gen/keptn-api-virtualservice.yaml
-verify_kubectl $? "Deploying keptn api virtualservice failed."
+  kubectl apply -f ../manifests/keptn/gen/keptn-api-virtualservice.yaml
+  verify_kubectl $? "Deploying keptn api virtualservice failed."
+fi
 
 helm init
 oc adm policy  add-cluster-role-to-user cluster-admin system:serviceaccount:kube-system:default
